pad: let FB_VAL() be a placeholder for optimized versions of fb_val()
This could be used to speed up fb_val() calls; doesn't matter much
for fbpad since glyphs are cached but for other framebuffer tools
that use draw.c this can be a real improvement.  See fb_val() function
to see how FB_VAL() should be implemented.  For instance, for a 32-bit
rgb framebuffer we can write:
  #define FB_VAL(r, g, b)	(((r) << 16) | ((g) << 8) | (b))