font: load font glyphs into memory
Loading glyphs from disk can result in noticeable delays in fbpad,
specially when the disk is busy.  What's more, modifying loaded font
files when fbpad is running may result in strange characters in fbpad.
On the other hand, fonts files are relatively small compared to the size
of framebuffers and very little memory can be saved by not loading them.

This patch changes fbpad to load the whole font file into the memory
to solve the mentioned problems.